combined tests into the test folder & updated some tests to be wrapped in test functions
