+++ test/test_datadir.py
