+++ test/test_exception_logging.py
